+++ engine/Cargo.toml
@@ -54,6 +54,7 @@ reearth-flow-action-processor = { path = "runtime/action-processor", default-fea
 reearth-flow-action-python-processor = { path = "runtime/action-python-processor" }
 reearth-flow-action-sink = { path = "runtime/action-sink" }
 reearth-flow-action-source = { path = "runtime/action-source" }
+reearth-flow-atlas = { path = "runtime/reearth-flow-atlas" }
 reearth-flow-common = { path = "runtime/common" }
 reearth-flow-eval-expr = { path = "runtime/eval-expr" }
 reearth-flow-geometry = { path = "runtime/geometry" }
